backend/requirements.txt
fastapi
uvicorn
boto3
python-dotenv
httpx
pyjwt